requirements.txt
Flask==2.3.2
tensorflow-cpu==2.18.0
Pillow==10.0.1
numpy==1.26.4
gunicorn==20.1.0
opencv-python-headless==4.9.0.80
mtcnn==0.1.1
